Restructure package layout
(Mostly) detangle GitHub-specific code from generic dependency graph implementation.
