+++ plugin/src/main/kotlin/org/gradle/dependencygraph/model/DependencyCoordinates.kt
@@ -1,0 +1,3 @@
+package org.gradle.dependencygraph.model
+
+data class DependencyCoordinates(val group: String, val module: String, val version: String)
