+++ plugin/src/main/kotlin/org/gradle/dependencygraph/model/ResolvedConfiguration.kt
@@ -1,0 +1,14 @@
+package org.gradle.dependencygraph.model
+
+data class ResolvedConfiguration(
+    val rootSource: DependencySource,
+    val allDependencies: MutableList<ResolvedDependency> = mutableListOf()
+) {
+    fun addDependency(component: ResolvedDependency) {
+        allDependencies.add(component)
+    }
+
+    fun hasDependency(componentId: String): Boolean {
+        return allDependencies.map { it.id }.contains(componentId)
+    }
+}
