+++ plugin/src/main/kotlin/org/gradle/github/GitHubDependencyGraphPlugin.kt
@@ -2,10 +2,7 @@ package org.gradle.github
 
 import org.gradle.api.Plugin
 import org.gradle.api.invocation.Gradle
-import org.gradle.api.logging.Logging
-import org.gradle.github.forceresolve.ForceDependencyResolutionPlugin
-import org.gradle.github.dependencygraph.GitHubDependencyExtractorPlugin
-import org.gradle.github.dependencygraph.internal.util.GradleExtensions
+import org.gradle.forceresolve.ForceDependencyResolutionPlugin
 
 /**
  * A plugin that collects all resolved dependencies in a Gradle build for submission to the
@@ -13,10 +10,6 @@ import org.gradle.github.dependencygraph.internal.util.GradleExtensions
  */
 @Suppress("unused")
 class GitHubDependencyGraphPlugin : Plugin<Gradle> {
-    private companion object : GradleExtensions() {
-        private val LOGGER = Logging.getLogger(GitHubDependencyGraphPlugin::class.java)
-    }
-
     override fun apply(gradle: Gradle) {
         // Only apply the dependency extractor to the root build
         if (gradle.parent == null) {
