+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/GitHubSnapshotParams.kt
@@ -1,0 +1,7 @@
+package org.gradle.github.dependencygraph
+
+class GitHubSnapshotParams(
+    val dependencyGraphJobCorrelator: String,
+    val dependencyGraphJobId: String,
+    val gitSha: String,
+    val gitRef: String)
