--- plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/ResolvedConfigurationFilter.kt
@@ -1,16 +1,0 @@
-package org.gradle.github.dependencygraph.internal
-
-class ResolvedConfigurationFilter(projectFilter: String?, configurationFilter: String?) {
-    val projectRegex = projectFilter?.toRegex()
-    val configurationRegex = configurationFilter?.toRegex()
-
-    fun include(projectPath: String, configurationName: String): Boolean {
-        if (projectRegex != null && !projectRegex.matches(projectPath)) {
-            return false
-        }
-        if (configurationRegex != null && !configurationRegex.matches(configurationName)) {
-            return false
-        }
-        return true
-    }
-}
