--- plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/github/GitHubSnapshotParams.kt
@@ -1,7 +1,0 @@
-package org.gradle.github.dependencygraph.internal.github
-
-class GitHubSnapshotParams(
-    val dependencyGraphJobCorrelator: String,
-    val dependencyGraphJobId: String,
-    val gitSha: String,
-    val gitRef: String)
