--- plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/github/json/GitHubDependency.kt
@@ -1,11 +1,0 @@
-package org.gradle.github.dependencygraph.internal.github.json
-
-data class GitHubDependency(
-    val package_url: String,
-    val relationship: Relationship,
-    val dependencies: List<String>
-) {
-    enum class Relationship {
-        indirect, direct
-    }
-}
