--- plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/github/json/GitHubJob.kt
@@ -1,6 +1,0 @@
-package org.gradle.github.dependencygraph.internal.github.json
-
-data class GitHubJob(
-    val id: String,
-    val correlator: String
-)
