--- plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/github/json/GitHubManifest.kt
@@ -1,7 +1,0 @@
-package org.gradle.github.dependencygraph.internal.github.json
-
-data class GitHubManifest(
-    val name: String,
-    val resolved: Map<String, GitHubDependency>,
-    val file: GitHubManifestFile?
-)
