--- plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/model/DependencyCoordinates.kt
@@ -1,3 +1,0 @@
-package org.gradle.github.dependencygraph.internal.model
-
-data class DependencyCoordinates(val group: String, val module: String, val version: String)
