+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/model/GitHubDependency.kt
@@ -1,0 +1,11 @@
+package org.gradle.github.dependencygraph.model
+
+data class GitHubDependency(
+    val package_url: String,
+    val relationship: Relationship,
+    val dependencies: List<String>
+) {
+    enum class Relationship {
+        indirect, direct
+    }
+}
