+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/model/GitHubDetector.kt
@@ -1,0 +1,7 @@
+package org.gradle.github.dependencygraph.model
+
+data class GitHubDetector(
+    val name: String = GitHubDetector::class.java.`package`.implementationTitle,
+    val version: String = GitHubDetector::class.java.`package`.implementationVersion,
+    val url: String = "https://github.com/gradle/github-dependency-graph-gradle-plugin"
+)
