+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/model/GitHubJob.kt
@@ -1,0 +1,6 @@
+package org.gradle.github.dependencygraph.model
+
+data class GitHubJob(
+    val id: String,
+    val correlator: String
+)
