+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/model/GitHubManifest.kt
@@ -1,0 +1,7 @@
+package org.gradle.github.dependencygraph.model
+
+data class GitHubManifest(
+    val name: String,
+    val resolved: Map<String, GitHubDependency>,
+    val file: GitHubManifestFile?
+)
