+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/model/GitHubManifestFile.kt
@@ -1,0 +1,8 @@
+package org.gradle.github.dependencygraph.model
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class GitHubManifestFile(
+    @get:JsonProperty("source_location")
+    val sourceLocation: String
+)
